Cleaner bug fix
Kwargs are used in the battle function to allow default arguments
but it seems that because the first argument was not a kwarg, the
player name was duplicated. After making the player argument a kwarg
and specifying each argument in each call of the battle function,
the #14 bug was fixed.